bootcamplist refactored
