+++ src/app/app.config.ts
@@ -3,8 +3,9 @@ import { provideRouter } from '@angular/router';
 
 import { routes } from './app.routes';
 import { provideHttpClient, withInterceptors } from '@angular/common/http';
-
 import { AuthInterceptor } from './core/interceptors/auth/auth.interceptor';
+
+
 
 
 export const appConfig: ApplicationConfig = {
