+++ src/app/core/guards/admin/admin-panel.guard.ts
@@ -3,7 +3,8 @@
 
 import { inject } from '@angular/core';
 import { CanActivateFn, ActivatedRouteSnapshot, RouterStateSnapshot, Router } from '@angular/router';
-import { AuthService } from '../../features/services/concretes/auth.service';
+import { AuthService } from '../../../features/services/concretes/auth.service';
+
 
 export const AdminPanelGuard: CanActivateFn = (route: ActivatedRouteSnapshot, state: RouterStateSnapshot) => {
   const authService = inject(AuthService);
