+++ src/app/features/components/bootcamps/bootcamp-list-group/bootcamp-list-group.component.html
@@ -1,30 +1,43 @@
-<app-instructor></app-instructor>
-<div class="row" style="margin-top: 10px;">
-    <h3 *ngIf="!bootcampList">bootcamp Bulunamadı</h3>
- <div *ngFor="let bootcamp of bootcampList.items" class="card" style="width: 16rem;margin:20px;opacity: 0.8;">
- <img [src]="bootcamp.imagePath" class="card-img-top" alt="arabanın adı Image" height="150px" />
- <div class="car-body">
-    <p class="card-text">Name:<strong>{{bootcamp.name}}</strong></p>
- </div>
- <ul class="list-group list-group-flush">
-  <li class="list-group-item"><b>name:</b><span style="float:right;">{{bootcamp.name}}</span></li>
-  <li class="list-group-item"><b>Hoca:</b><span style="float:right;">{{bootcamp.instructorFirstName + " " + bootcamp.instructorLastName}}</span></li>
-  <li class="list-group-item"><b>startDate:</b><span style="float:right;">{{bootcamp.startDate}}</span></li>
-  <li class="list-group-item"><b>endDate:</b><span style="float:right;">{{bootcamp.endDate}}</span></li>
-  
-  <!-- <button *ngIf="isBeforeEnd(bootcamp.endDate)">Başvur</button> -->
-  <!-- <button class="text-white bg-blue-700 hover:bg-blue-800 focus:outline-none focus:ring-4 focus:ring-blue-300 font-medium rounded-full text-sm px-5 py-2.5 text-center me-2 mb-2 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800" [ngClass]="{'expired': isExpired(bootcamp.endDate), 'notExpired': !isExpired(bootcamp.endDate)}"><a routerLink="/bootcamps/bootcamp.id">Detaya Git</a> </button> -->
-  <button class="text-white bg-blue-700 hover:bg-blue-800 focus:outline-none focus:ring-4 focus:ring-blue-300 font-medium rounded-full text-sm px-5 py-2.5 text-center me-2 mb-2 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800"
-  [ngClass]="{'expired': isExpired(bootcamp.endDate), 'notExpired': !isExpired(bootcamp.endDate)}"
-  [routerLink]="['/bootcampdetail', bootcamp.id.toString()]">
-Detaya Git
-</button>
-</ul>
-</div> 
-<button *ngIf="bootcampList.hasNext" type="button"
-     (click)="onViewMoreClicked()" [disabled]="!bootcampList.hasNext" class="btn border-0">
-     Sonraki Sayfa {{this.currentPageNumber}} {{bootcampList.hasNext}}</button>
-     <button *ngIf="bootcampList.hasPrevious" type="button"
-        (click)="onPreviousPageClicked()" [disabled]="!bootcampList.hasPrevious" class="btn border-0">
-        Önceki Sayfa {{this.currentPageNumber-1}} {{bootcampList.hasPrevious}}</button>
+<div class="flex h-screen bg-gray-100">
+   <!-- Sidebar -->
+   <aside class="w-64 bg-white shadow-md">
+       <div class="p-4">
+           <app-instructor></app-instructor>
+       </div>
+   </aside>
+   
+   <!-- Content -->
+   <div class="flex-grow p-8">
+       <div class="container mx-auto">
+           <h3 *ngIf="!bootcampList" class="text-center">Bootcamp Bulunamadı</h3>
+           
+           <div class="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-6 justify-center">
+               <div *ngFor="let bootcamp of bootcampList.items" class="max-w-sm rounded overflow-hidden shadow-lg">
+                   <img [src]="bootcamp.imagePath" class="w-full h-48 object-cover" alt="arabanın adı Image">
+                   <div class="px-6 py-4">
+                       <div class="font-bold text-xl mb-2">{{bootcamp.name}}</div>
+                       <p class="text-gray-700 text-base"><b>Hoca:</b> {{bootcamp.instructorFirstName}} {{bootcamp.instructorLastName}}</p>
+                       <p class="text-gray-700 text-base"><b>Start Date:</b> {{bootcamp.startDate}}</p>
+                       <p class="text-gray-700 text-base"><b>End Date:</b> {{bootcamp.endDate}}</p>
+                   </div>
+                   <div class="px-6 py-4">
+                       <button class="text-white bg-blue-700 hover:bg-blue-800 focus:outline-none focus:ring-4 focus:ring-blue-300 font-medium rounded-full text-sm px-5 py-2.5 text-center"[ngClass]="{'expired': isExpired(bootcamp.endDate), 'notExpired': !isExpired(bootcamp.endDate)}"
+                       [routerLink]="['/bootcampdetail', bootcamp.id.toString()]">
+                           Detaya Git
+                       </button>
+                   </div>
+               </div>
+           </div>
+           
+           <div class="flex justify-center mt-4">
+               <button *ngIf="bootcampList.hasNext" type="button" (click)="onViewMoreClicked()" [disabled]="!bootcampList.hasNext" class="btn border-0 mr-2">
+                   Sonraki Sayfa {{this.currentPageNumber}} {{bootcampList.hasNext}}
+               </button>
+               
+               <button *ngIf="bootcampList.hasPrevious" type="button" (click)="onPreviousPageClicked()" [disabled]="!bootcampList.hasPrevious" class="btn border-0">
+                   Önceki Sayfa {{this.currentPageNumber-1}} {{bootcampList.hasPrevious}}
+               </button>
+           </div>
+       </div>
+   </div>
 </div>
